Fixed the player Interact + UI
Added a Interface for interact that scripts who need to be interacted with can inherit. This should make things smoother in the future.

The interaction is now handled by the player.

Updated the door trigger so it now uses the new interact + can handle multiple doors.

Added so the pickup can be interacted with.